More moving & renaming files
No reason the job runner can only run in AWE - it should have no idea
where it's running.
