+++ src/us/kbase/narrativejobservice/NarrativeJobServiceServer.java
@@ -43,6 +43,7 @@ import us.kbase.common.taskqueue2.TaskQueue;
 import us.kbase.common.taskqueue2.TaskQueueConfig;
 import us.kbase.narrativejobservice.db.ExecEngineMongoDb;
 import us.kbase.narrativejobservice.db.MigrationToMongo;
+import us.kbase.narrativejobservice.sdkjobs.ErrorLogger;
 import us.kbase.userandjobstate.InitProgress;
 import us.kbase.userandjobstate.Results;
 import us.kbase.userandjobstate.UserAndJobStateClient;
