+++ src/us/kbase/narrativejobservice/RunAppBuilder.java
@@ -56,6 +56,7 @@ import us.kbase.narrativejobservice.db.ExecEngineMongoDb;
 import us.kbase.narrativejobservice.db.ExecLog;
 import us.kbase.narrativejobservice.db.ExecLogLine;
 import us.kbase.narrativejobservice.db.ExecTask;
+import us.kbase.narrativejobservice.sdkjobs.ErrorLogger;
 import us.kbase.shock.client.BasicShockClient;
 import us.kbase.shock.client.ShockACLType;
 import us.kbase.shock.client.ShockNodeId;
