+++ src/us/kbase/narrativejobservice/Util.java
@@ -17,16 +17,18 @@ import us.kbase.common.service.UObject;
 import us.kbase.userandjobstate.UserAndJobStateClient;
 
 public class Util {
+	
+	// used in RemoteAppTester & RunAppTest test classes
 	public static void waitForJob(String token, String ujsUrl, String jobId) throws Exception {
 		UserAndJobStateClient jscl = new UserAndJobStateClient(new URL(ujsUrl), new AuthToken(token));
 		jscl.setAllSSLCertificatesTrusted(true);
 		jscl.setIsInsecureHttpConnectionAllowed(true);
 		for (@SuppressWarnings("unused")
-        int iter = 0; ; iter++) {
+		int iter = 0; ; iter++) {
 			Thread.sleep(5000);
 			Tuple7<String, String, String, Long, String, Long, Long> data = jscl.getJobStatus(jobId);
-    		Long complete = data.getE6();
-    		Long wasError = data.getE7();
+			Long complete = data.getE6();
+			Long wasError = data.getE7();
 			//System.out.println("Status (" + iter + "): " + data.getE3());
 			if (complete == 1L) {
 				if (wasError == 0L) {
@@ -41,6 +43,7 @@ public class Util {
 		}
 	}
 	
+	// used in the NJS *App* methods, not in SDK methods
 	public static boolean isAweJobId(String jobId) {
 		// xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx
 		try {
@@ -51,32 +54,34 @@ public class Util {
 		}
 	}
 	
+	// currently unused in the NJS wrapper codebase
 	public static boolean isUjsJobId(String jobId) {
 		// 545a7b6ee4b0d82af0eafa16
 		return jobId.length() == 24;
 	}
 	
-    @SuppressWarnings("unchecked")
-    public static String addShockNodePublicReadACL(String shockUrl, String token, 
-            String shockNodeId) throws Exception {
-        String nodeurl = shockUrl;
-        if (!nodeurl.endsWith("/"))
-            nodeurl += "/";
-        nodeurl += "node/" + shockNodeId + "/acl/public_read";
-        final HttpPut htp = new HttpPut(nodeurl);
-        PoolingHttpClientConnectionManager cm = new PoolingHttpClientConnectionManager();
-        cm.setMaxTotal(1000);
-        cm.setDefaultMaxPerRoute(1000);
-        CloseableHttpClient client = HttpClients.custom().setConnectionManager(cm).build();
-        htp.setHeader("Authorization", "OAuth " + token);
-        final CloseableHttpResponse response = client.execute(htp);
-        try {
-            String resp = EntityUtils.toString(response.getEntity());
-            Map<String, String> node = (Map<String, String>)UObject.getMapper()
-                    .readValue(resp, Map.class).get("data");
-            return node.get("id");
-        } finally {
-            response.close();
-        }
-    }
+	// currently unused in the NJS wrapper codebase
+	@SuppressWarnings("unchecked")
+	public static String addShockNodePublicReadACL(String shockUrl, String token, 
+			String shockNodeId) throws Exception {
+		String nodeurl = shockUrl;
+		if (!nodeurl.endsWith("/"))
+			nodeurl += "/";
+		nodeurl += "node/" + shockNodeId + "/acl/public_read";
+		final HttpPut htp = new HttpPut(nodeurl);
+		PoolingHttpClientConnectionManager cm = new PoolingHttpClientConnectionManager();
+		cm.setMaxTotal(1000);
+		cm.setDefaultMaxPerRoute(1000);
+		CloseableHttpClient client = HttpClients.custom().setConnectionManager(cm).build();
+		htp.setHeader("Authorization", "OAuth " + token);
+		final CloseableHttpResponse response = client.execute(htp);
+		try {
+			String resp = EntityUtils.toString(response.getEntity());
+			Map<String, String> node = (Map<String, String>)UObject.getMapper()
+					.readValue(resp, Map.class).get("data");
+			return node.get("id");
+		} finally {
+			response.close();
+		}
+	}
 }
